move span utils to new file
